refactor(react-components): プリミティブをフォルダコロケーション化
primitive/{name}.tsx を primitive/{name}/{name}.tsx に移動し、
本体・ストーリー・バレル（index.ts）の3点セットでまとめる。
view/{view-name}/ と同じ構造になり、サブコンポーネントの追加や
昇格時の構造的な対称性が保たれる。

dialog は modal の内部実装としてのみ利用されており、ADR-024 の
「特定ビュー専用のサブコンポーネントは最初に使うビューのディレクトリ
配下にコロケーション」と同じ判断で modal/dialog.tsx に移し、
modal/index.ts からは再エクスポートしない（internal 扱い）。

primitive/index.ts はフォルダ解決で各 {name}/index.ts を拾うため
変更不要。利用側 import パスにも影響なし。
