+++ packages/react-components/src/primitive/button/index.ts
@@ -1,0 +1,1 @@
+export * from "./button";
