+++ packages/react-components/src/primitive/error-alert/index.ts
@@ -1,0 +1,1 @@
+export * from "./error-alert";
