+++ packages/react-components/src/primitive/gym-context/index.ts
@@ -1,0 +1,1 @@
+export * from "./gym-context";
