+++ packages/react-components/src/primitive/link/index.ts
@@ -1,0 +1,1 @@
+export * from "./link";
