+++ packages/react-components/src/primitive/modal/index.ts
@@ -1,0 +1,1 @@
+export * from "./modal";
