+++ packages/react-components/src/primitive/modal/modal.stories.tsx
@@ -2,7 +2,7 @@
 
 import type { Meta, StoryObj } from "@storybook/react";
 import { expect, userEvent, within } from "storybook/test";
-import { Button } from "./button";
+import { Button } from "../button";
 import {
 	Modal,
 	ModalBody,
