+++ packages/react-components/src/primitive/page-heading/index.ts
@@ -1,0 +1,1 @@
+export * from "./page-heading";
