+++ packages/react-components/src/primitive/page-section/index.ts
@@ -1,0 +1,1 @@
+export * from "./page-section";
